Move theme out of shared dir
