+++ fronts-client/src/shared/components/Fadeable.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 
 export default styled.div<{
   fade?: boolean;
