+++ fronts-client/src/shared/components/Link.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 
 export default styled.a.attrs({
   target: '_blank',
