+++ fronts-client/src/shared/components/alert/InformationMsg.tsx
@@ -1,5 +1,5 @@
-import { styled } from 'shared/constants/theme';
-import { theme } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
+import { theme } from 'constants/theme-shared';
 
 export default styled.div<{ status?: 'info' | 'error' }>`
   position: relative;
