+++ fronts-client/src/shared/components/article/DraggableArticleImageContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 import { connect } from 'react-redux';
 import { State } from 'types/State';
 import {
