+++ fronts-client/src/shared/components/card/CardDraftMetaContent.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 import CardMetaContent from './CardMetaContent';
 import { notLiveColour } from 'shared/util/getPillarColor';
