+++ fronts-client/src/shared/components/collection/CollectionMetaContainer.tsx
@@ -1,4 +1,4 @@
-import { styled } from '../../constants/theme';
+import { styled } from '../../../constants/theme-shared';
 
 export default styled.div`
   display: flex;
