+++ fronts-client/src/shared/components/icons/HoverIcons.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { theme } from 'shared/constants/theme';
+import { theme } from 'constants/theme-shared';
 interface IconProps {
   fill?: string;
   size?: number;
