+++ fronts-client/src/shared/components/input/InputBase.ts
@@ -1,4 +1,4 @@
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 
 export default styled.input<{
   useHeadlineFont?: boolean;
