+++ fronts-client/src/shared/components/input/InputContainer.ts
@@ -1,6 +1,6 @@
 // @flow
 
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 
 // This component has no styles - we don't have a style for input containers,
 // but other container components need to refer to them, for example
