+++ fronts-client/src/shared/components/input/InputGroup.ts
@@ -1,4 +1,4 @@
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 import InputContainer from './InputContainer';
 
 export default styled.div`
