+++ fronts-client/src/shared/components/input/InputLabel.ts
@@ -1,4 +1,4 @@
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 
 export default styled.label<{
   size?: 'sm';
