+++ fronts-client/src/shared/components/layout/HorizontalRule.ts
@@ -1,4 +1,4 @@
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 
 export default styled.div<{ noMargin?: boolean }>`
   width: 100%;
