+++ fronts-client/src/shared/components/snapLink/SnapLink.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { styled } from 'shared/constants/theme';
+import { styled } from 'constants/theme-shared';
 import { connect } from 'react-redux';
 import upperFirst from 'lodash/upperFirst';
